pyproject.toml
[tool.poetry]
name = "themefinder"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "themefinder", from = "src"}]

[tool.poetry.dependencies]
langchain = "*"
langchain-openai = "0.1.17"
python = ">=3.12,<3.13"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-asyncio = "^0.24.0"
langfuse = "2.29.1"
boto3 = "^1.29"
scikit-learn = "*"
openpyxl = "^3.1.5"
pyarrow = "^15.0.0" 

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
